+++ python/sdssdb/connection.py
@@ -13,7 +13,6 @@ import importlib
 import os
 import re
 import socket
-import warnings
 
 import pgpasslib
 import six
@@ -25,28 +24,16 @@ from sqlalchemy.ext.declarative import DeferredReflection
 from sqlalchemy.orm import scoped_session, sessionmaker
 
 import peewee
-from peewee import OperationalError
+from peewee import OperationalError, PostgresqlDatabase
 from playhouse.postgres_ext import ArrayField
 from playhouse.reflection import Introspector, UnknownField
 
 import sdssdb
-from sdssdb import config, log, use_psycopg3
+from sdssdb import config, log
 from sdssdb.utils.internals import get_database_columns
 
 
 __all__ = ["DatabaseConnection", "PeeweeDatabaseConnection", "SQLADatabaseConnection"]
-
-
-if use_psycopg3 is True:
-    try:
-        import psycopg  # noqa
-        from playhouse.psycopg3_ext import Psycopg3Database as PostgresqlDatabase
-    except ImportError:
-        warnings.warn("psycopg3 is not installed. Using psycopg2.")
-        from peewee import PostgresqlDatabase
-
-else:
-    from peewee import PostgresqlDatabase
 
 
 def _should_autoconnect():
@@ -229,7 +216,8 @@ class DatabaseConnection(six.with_metaclass(abc.ABCMeta)):
         Parameters
         ----------
         dbname : `str` or `None`
-            The database name. If `None`, defaults to `.dbname`.
+            The database name. If `None`, defaults to `.dbname`. ``dbname`` can also be
+            a full database URI, in which case the other connection parameters are ignored.
         user : str
             Overrides the profile database user.
         host : str
@@ -272,7 +260,9 @@ class DatabaseConnection(six.with_metaclass(abc.ABCMeta)):
             )
 
         return self.connect_from_parameters(
-            dbname=dbname, silent_on_fail=silent_on_fail, **db_configuration
+            dbname=dbname,
+            silent_on_fail=silent_on_fail,
+            **db_configuration,
         )
 
     def connect_from_parameters(self, dbname=None, **params):
@@ -332,13 +322,21 @@ class DatabaseConnection(six.with_metaclass(abc.ABCMeta)):
         """Returns the URI to the database connection."""
 
         params = self.connection_params
-        if not self.connected or params is None:
+        if not self.connected or params is None or self.dbname is None:
             raise RuntimeError("The database is not connected.")
 
-        return get_database_uri(self.dbname, **params)
+        valid_params = {
+            "user": params.get("user", None),
+            "host": params.get("host", None),
+            "port": params.get("port", None),
+            "password": params.get("password", None),
+        }
 
+        return get_database_uri(self.dbname, **valid_params)
+
+    @property
     @abc.abstractmethod
-    def connection_params(self):
+    def connection_params(self) -> dict | None:
         """Returns a dictionary with the connection parameters.
 
         Returns
@@ -355,10 +353,11 @@ class DatabaseConnection(six.with_metaclass(abc.ABCMeta)):
     def become(self, user):
         """Change the connection to a certain user."""
 
-        if not self.connected:
+        dsn_params = self.connection_params
+
+        if not self.connected or dsn_params is None:
             raise RuntimeError("DB has not been initialised.")
 
-        dsn_params = self.connection_params
         dsn_params.pop("password", None)  # Do not keep the password since it may change.
 
         if dsn_params is None:
@@ -430,13 +429,16 @@ class PeeweeDatabaseConnection(DatabaseConnection, PostgresqlDatabase):
 
     """
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args, use_psycopg3=None, **kwargs):
         self.models = {}
         self.introspector = {}
 
         self._metadata = {}
 
-        PostgresqlDatabase.__init__(self, None)
+        if use_psycopg3 is not None:
+            sdssdb.use_psycopg3 = use_psycopg3
+
+        PostgresqlDatabase.__init__(self, None, prefer_psycopg3=sdssdb.use_psycopg3)
         DatabaseConnection.__init__(self, *args, **kwargs)
 
     @property
@@ -450,29 +452,60 @@ class PeeweeDatabaseConnection(DatabaseConnection, PostgresqlDatabase):
         """Returns a dictionary with the connection parameters."""
 
         if self.connected:
-            return self.connect_params.copy()
+            if self.psycopg_version == "psycopg2":
+                return self.connection().info.dsn_parameters
+            elif self.psycopg_version == "psycopg3":
+                return self.connection().info.get_parameters()
+            else:
+                raise RuntimeError("unknown psycopg version in use.")
 
         return None
+
+    @property
+    def psycopg_version(self):
+        """Returns the version of psycopg in use."""
+
+        if not self.connected:
+            raise RuntimeError("The database is not connected.")
+
+        if isinstance(self._adapter, self.psycopg2_adapter):
+            return "psycopg2"
+        elif isinstance(self._adapter, self.psycopg3_adapter):
+            return "psycopg3"
+        else:
+            return "unknown"
 
     def _conn(self, dbname, silent_on_fail=False, **params):
         """Connects to the DB and tests the connection."""
 
-        if "password" not in params:
-            pgpass_params = {
-                key: value for key, value in params.copy().items() if value is not None
-            }
+        if dbname.startswith("postgresql://"):
+            PostgresqlDatabase.__init__(self, dbname, prefer_psycopg3=sdssdb.use_psycopg3)
+        else:
+            if "password" not in params:
+                pgpass_params = {
+                    key: value for key, value in params.copy().items() if value is not None
+                }
 
-            try:
-                params["password"] = pgpasslib.getpass(dbname=dbname, **pgpass_params)
-            except pgpasslib.FileNotFound:
-                params["password"] = None
+                try:
+                    params["password"] = pgpasslib.getpass(dbname=dbname, **pgpass_params)
+                except pgpasslib.FileNotFound:
+                    params["password"] = None
 
-        PostgresqlDatabase.init(self, dbname, **params)
-        self._metadata = {}
+            PostgresqlDatabase.init(
+                self,
+                dbname,
+                prefer_psycopg3=sdssdb.use_psycopg3,
+                **params,
+            )
+            self._metadata = {}
 
         try:
             PostgresqlDatabase.connect(self)
+
+            conn_params = self.connection_params
+            dbname = conn_params.get("dbname", dbname)
             self.dbname = dbname
+
         except OperationalError as ee:
             if not silent_on_fail:
                 log.warning(f"failed connecting to database {self.database!r}: {ee}")
@@ -641,8 +674,11 @@ class SQLADatabaseConnection(DatabaseConnection):
 
         """
 
+        if dbname.startswith("postgresql://") or dbname.startswith("postgresql+psycopg"):
+            return dbname
+
         db_params = params.copy()
-        db_params["drivername"] = "postgresql+psycopg2"
+        db_params["drivername"] = "postgresql+psycopg"
         db_params["database"] = dbname
         db_params["username"] = db_params.pop("user", None)
         db_params["host"] = db_params.pop("host", "localhost")