+++ tests/test_connection.py
@@ -11,7 +11,11 @@
 # Modified By: Brian Cherinka
 
 
-from __future__ import print_function, division, absolute_import
+from __future__ import absolute_import, division, print_function
+
+import sdssdb
+import sdssdb.connection
+from sdssdb.connection import PeeweeDatabaseConnection
 
 
 def assert_testdb(database):
@@ -46,3 +50,21 @@ class TestGenericDatabaseConnection(object):
         # change back
         database.change_version()
         assert_testdb(database)
+
+
+def test_db_uri_peewee_psycopg2():
+    sdssdb.use_psycopg3 = False
+
+    database = PeeweeDatabaseConnection("postgresql://localhost/test")
+    assert database.dbname == "test"
+    assert database.connected is True
+    assert database.psycopg_version == "psycopg2"
+
+
+def test_db_uri_peewee_psycopg3():
+    sdssdb.use_psycopg3 = True
+
+    database = PeeweeDatabaseConnection("postgresql://localhost/test")
+    assert database.dbname == "test"
+    assert database.connected is True
+    assert database.psycopg_version == "psycopg3"